Applying 4.2.1 room availability fix to 4.3
An issue arose in 4.2.1 whilst browsing in Safari where the room
priceFrom is being populated in scope, even when the room is
unavailable.

This prevented the “Not available on selected dates” message from
displaying.

I have removed the check for priceFrom when deciding whether to display
this message because priceFrom is irrelevant. We should only be
interested in whether the room is available or not.