Changing funnel retention search in session checks
Adding new cookie to denote if a user has searched in their current
session.

Retention now checks against this rather than if the user has any
previous searches.

This means that retention fires if the user has not carried out
searches in the current session, not if the user has carried out
searches at some point in the past